Reusing existing getStripeElementOptions function
