+++ client/blocks/utils.js
@@ -1,5 +1,8 @@
 /* global wc */
 
+import { isLinkEnabled } from 'wcstripe/stripe-utils';
+import { OPTIMIZED_CHECKOUT_DEFAULT_LAYOUT } from 'wcstripe/stripe-utils/constants';
+
 export const getBlocksConfiguration = () => {
 	const stripeServerData = wc?.wcSettings?.getSetting( 'stripe_data', null );
 
@@ -141,4 +144,73 @@ export const addOrderAttributionInputsIfNotExists = () => {
 export const getStripeImageUrl = ( imageName ) => {
 	const config = getBlocksConfiguration();
 	return `${ config?.plugin_url }/assets/images/${ imageName }.svg`;
+};
+
+/**
+ * Gets the Stripe element options.
+ *
+ * @return {Object} The Stripe element options.
+ */
+export const getStripeElementOptions = () => {
+	let options = {
+		fields: {
+			billingDetails: {
+				name: 'never',
+				email: 'never',
+				// The phone field is optional, so it needs to be "auto" to not throw errors
+				// when passing the phone parameter to create a payment method.
+				phone: 'auto',
+				address: {
+					country: 'never',
+					line1: 'never',
+					line2: 'never',
+					city: 'never',
+					state: 'never',
+					postalCode: 'never',
+				},
+			},
+		},
+		wallets: {
+			applePay: 'never',
+			googlePay: 'never',
+		},
+	};
+
+	// Prefill Link customer data if available.
+	if ( isLinkEnabled() ) {
+		const userEmail = document.getElementById( 'email' )?.value;
+		if ( userEmail ) {
+			const userPhone =
+				document.getElementById( 'billing-phone' )?.value ||
+				document.getElementById( 'shipping-phone' )?.value;
+
+			options = {
+				...options,
+				defaultValues: {
+					billingDetails: {
+						email: userEmail,
+						phone: userPhone,
+					},
+				},
+			};
+		}
+	}
+
+	if ( getBlocksConfiguration()?.isOCEnabled ) {
+		const layout = {
+			type:
+				getBlocksConfiguration()?.OCLayout ||
+				OPTIMIZED_CHECKOUT_DEFAULT_LAYOUT,
+		};
+		if ( layout.type === OPTIMIZED_CHECKOUT_DEFAULT_LAYOUT ) {
+			layout.radios = false;
+			layout.spacedAccordionItems = false;
+		}
+		options = {
+			...options,
+			layout,
+		};
+	}
+
+	return options;
 };
